+++ app/src/main/java/com/example/myratp/ui/timetable/metrolines/MetroSchedulesActivity.kt
@@ -1,7 +1,14 @@
 package com.example.myratp.ui.timetable.metrolines
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
+import android.view.View
+import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -11,6 +18,9 @@ import com.example.myratp.adapters.MetroScheduleAdapter
 import com.example.myratp.data.AppDatabase
 import com.example.myratp.data.ScheduleDao
 import com.example.myratp.model.Schedule
+import kotlinx.android.synthetic.main.activity_bus_time.*
+import kotlinx.android.synthetic.main.activity_bus_time.progress_bar
+import kotlinx.android.synthetic.main.activity_metro_schedule.*
 import kotlinx.coroutines.runBlocking
 
 class MetroSchedulesActivity : AppCompatActivity(){
@@ -19,6 +29,7 @@ class MetroSchedulesActivity : AppCompatActivity(){
     private var name : String? = ""
     private var scheduleDao : ScheduleDao? = null
 
+    @RequiresApi(Build.VERSION_CODES.M)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_metro_schedule)
@@ -35,20 +46,43 @@ class MetroSchedulesActivity : AppCompatActivity(){
             .build()
         scheduleDao = database.getScheduleDao()
 
-        runBlocking {
-            scheduleDao?.deleteAllSchedule()
-            val service = retrofit().create(MetroLinesBySearch::class.java)
-            val resultat = service.getScheduleMetro("metros", "$code", "$name", "A+R")
-            resultat.result.schedules.map {
-                val metroSchedule = Schedule(0, it.message, it.destination)
-                Log.d("CCC", "$metroSchedule")
-                scheduleDao?.addSchedule(metroSchedule)
+        if(isNetworkConnected()){
+            runBlocking {
+                scheduleDao?.deleteAllSchedule()
+                val service = retrofit().create(MetroLinesBySearch::class.java)
+                val resultat = service.getScheduleMetro("metros", "$code", "$name", "A+R")
+                resultat.result.schedules.map {
+                    val metroSchedule = Schedule(0, it.message, it.destination)
+                    Log.d("CCC", "$metroSchedule")
+                    scheduleDao?.addSchedule(metroSchedule)
+                }
+                scheduleDao = database.getScheduleDao()
+                val s = scheduleDao?.getSchedule()
+                progress_bar.visibility = View.GONE
+                recyclerview_metro_schedule.adapter =
+                    MetroScheduleAdapter(s ?: emptyList())
             }
-            scheduleDao = database.getScheduleDao()
-            val s = scheduleDao?.getSchedule()
-            //val test = s.isNullOrEmpty()
-            recyclerview_metro_schedule.adapter =
-                MetroScheduleAdapter(s ?: emptyList())
         }
+        else{
+            Toast.makeText(this, "Vérifiez votre connexion internet et réessayez à nouveau", Toast.LENGTH_SHORT).show()
+        }
+
+
+    }
+
+    @RequiresApi(Build.VERSION_CODES.M)
+    private fun isNetworkConnected(): Boolean {
+        var result = false
+        val connectivityManager = applicationContext.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager?
+        connectivityManager?.let{
+            it.getNetworkCapabilities(connectivityManager.activeNetwork)?.apply {
+                result = when {
+                    hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                    hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                    else -> false
+                }
+            }
+        }
+        return result
     }
 }
