+++ app/src/main/java/com/example/myratp/ui/timetable/metrolines/MetroTimeActivity.kt
@@ -1,8 +1,15 @@
 package com.example.myratp.ui.timetable.metrolines
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.view.View
+import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.room.Room
@@ -13,6 +20,9 @@ import com.example.myratp.data.MetroLineDao
 import com.example.myratp.data.TrafficDao
 import com.example.myratp.model.MetroLine
 import com.example.myratp.model.Traffic
+import kotlinx.android.synthetic.main.activity_bus_time.*
+import kotlinx.android.synthetic.main.activity_bus_time.progress_bar
+import kotlinx.android.synthetic.main.activity_metro_time.*
 import kotlinx.coroutines.runBlocking
 
 class MetroTimeActivity : AppCompatActivity() {
@@ -20,6 +30,7 @@ class MetroTimeActivity : AppCompatActivity() {
     private var metroLineDao : MetroLineDao? = null
     private var trafficDao : TrafficDao? = null
 
+    @RequiresApi(Build.VERSION_CODES.M)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_metro_time)
@@ -35,41 +46,56 @@ class MetroTimeActivity : AppCompatActivity() {
             .build()
         trafficDao = database_bis.getTrafficDao()
 
-//        runBlocking {
-//
-//        }
+        if(isNetworkConnected()){
+            runBlocking {
 
-        runBlocking {
+                trafficDao?.deleteAllTraffic()
+                val service_bis = retrofit().create(MetroLinesBySearch::class.java)
+                val resultat_bis = service_bis.getTrafficMetro("metros")
+                resultat_bis.result.metros.map {
+                    val traffic = Traffic(0, it.line, it.slug, it.title, it.message)
+                    trafficDao?.addTraffic(traffic)
+                }
+                trafficDao = database_bis.getTrafficDao()
+                val traf = trafficDao?.getTraffic()
 
-            trafficDao?.deleteAllTraffic()
-            val service_bis = retrofit().create(MetroLinesBySearch::class.java)
-            val resultat_bis = service_bis.getTrafficMetro("metros")
-            resultat_bis.result.metros.map {
-                val traffic = Traffic(0, it.line, it.slug, it.title, it.message)
-                trafficDao?.addTraffic(traffic)
+
+                metroLineDao?.deleteAllMetroLines()
+                val service = retrofit().create(MetroLinesBySearch::class.java)
+                val resultat = service.getlistMetroLine()
+                resultat.result.metros.map {
+                    val metro = MetroLine(0, it.code, it.name, it.directions, it.id)
+                    Log.d("CCC", "$metro")
+                    metroLineDao?.addMetroLines(metro)
+                }
+                metroLineDao = database.getMetroLineDao()
+                val ms = metroLineDao?.getMetroLines()
+                progress_bar.visibility = View.GONE
+                recyclerview_metro.adapter =
+                    MetroLineAdapter(ms ?: emptyList(), traf!!)
             }
-            trafficDao = database_bis.getTrafficDao()
-            val traf = trafficDao?.getTraffic()
-            //val test = bs.isNullOrEmpty()
-//            recyclerview_metro.adapter =
-//                MetroLineAdapter(traf ?: emptyList())
-
-
-            metroLineDao?.deleteAllMetroLines()
-            val service = retrofit().create(MetroLinesBySearch::class.java)
-            val resultat = service.getlistMetroLine()
-            resultat.result.metros.map {
-                val metro = MetroLine(0, it.code, it.name, it.directions, it.id)
-                Log.d("CCC", "$metro")
-                metroLineDao?.addMetroLines(metro)
-            }
-            metroLineDao = database.getMetroLineDao()
-            val ms = metroLineDao?.getMetroLines()
-            //val test = bs.isNullOrEmpty()
-            recyclerview_metro.adapter =
-                MetroLineAdapter(ms ?: emptyList(), traf!!)
+        }
+        else{
+            Toast.makeText(this, "Vérifiez votre connexion internet et réessayez à nouveau", Toast.LENGTH_SHORT).show()
         }
 
 
     }
+
+    @RequiresApi(Build.VERSION_CODES.M)
+    private fun isNetworkConnected(): Boolean {
+        var result = false
+        val connectivityManager = applicationContext.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager?
+        connectivityManager?.let{
+            it.getNetworkCapabilities(connectivityManager.activeNetwork)?.apply {
+                result = when {
+                    hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                    hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                    else -> false
+                }
+            }
+        }
+        return result
+    }
+
 }
